split server code and add cors
